% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{aggregate_feature}
\alias{aggregate_feature}
\title{aggregate value per feature}
\usage{
aggregate_feature(annomatrix, columnstoaggregate, aggregatebycolumn, operation)
}
\arguments{
\item{annomatrix}{dataframe}

\item{columnstoaggregate}{column to use for aggreagtion, eg. values}

\item{aggregatebycolumn}{column will be by use for group by process eg. gene}

\item{operation}{eg. mean (not under the quotes)}
}
\value{
Dataframe of aggregated values by operation of interest
}
\description{
aggregate value per feature
}
\examples{
agg <- aggregate_feature(atacseqkd_quantify_list$featurecounts$histonemarks$summed$merged_diff_table_afr_st, c(3), "id", mean)
}
