% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{chipseeker_region_anno}
\alias{chipseeker_region_anno}
\title{chipseeker run and annotation}
\usage{
chipseeker_region_anno(
  assaytype,
  software,
  contrasts,
  region_bed,
  path,
  txdb,
  org_db,
  regionformat,
  fdr,
  fc
)
}
\arguments{
\item{assaytype}{eg. atacseq}

\item{software}{eg. deseq2}

\item{contrasts}{contrasts list}

\item{path}{location of peaks files (eg. .bed , .narrowPeak, .broadPeak)}

\item{txdb}{eg. TxDb.Hsapiens.UCSC.hg38.knownGene (not quoted)}

\item{org_db}{eg. org.Hs.eg.db}

\item{regionformat}{eg. ".bed", all peak files should have same format}

\item{fdr}{p.adj-value threshold}

\item{fc}{fold change threshold}
}
\value{
list containing objects, granges, plots from chipseeker
}
\description{
chipseeker run and annotation
}
\examples{
annotation <- chipseeker_region_anno("atacseqkd", "deseq2", atacseqkd_deseq2_list$dar_analysis$de_analysis, atacseqkd_deseq2_list$consensus[,c(1:4)], "/mnt/home3/reid/av638/atacseq/iva_lab_gencode/boutfolder/", TxDb.Hsapiens.UCSC.hg38.knownGene, "org.Hs.eg.db", "region_bed.txt", 0.05, 2)
}
