% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{consensus_gene_anno}
\alias{consensus_gene_anno}
\title{diffbind consensus gene annotation}
\usage{
consensus_gene_anno(
  assaytype,
  software,
  contrasts,
  consensus_bed,
  genefile,
  path,
  outformat,
  fdr,
  fc
)
}
\arguments{
\item{assaytype}{eg. atacseq, rnaseq}

\item{software}{internal software to be used by Diffbind eg. edger or deseq2}

\item{contrasts}{contrasts list}

\item{genefile}{gene file name}

\item{path}{location of files}

\item{outformat}{eg. ".txt"}

\item{fdr}{p.adj-value threshold}

\item{fc}{fold change threshold}
}
\value{
list of annotated differential regions consensus
}
\description{
consensus peaks from nextflow output but can be applied to other sets
}
\examples{
annotation <- consensus_gene_anno("atacseqkd", "deseq2", atacseqkd_deseq2_list$dar_analysis$de_analysis, atacseqkd_deseq2_list$consensus, atacseqkd_deseq2_list$gene, "/mnt/home3/reid/av638/atacseq/iva_lab_gencode/boutfolder", "txt", 0.05, 2)
}
