% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{consensus_gene_count_anno}
\alias{consensus_gene_count_anno}
\title{diffbind consensus gene annotation for normalized count matrix (consensus peaks from nextflow output but can be applied to other sets)}
\usage{
consensus_gene_count_anno(
  assaytype,
  software,
  normcountmatrix,
  consensus_bed,
  genefile,
  path,
  outformat
)
}
\arguments{
\item{assaytype}{eg. atacseq, rnaseq}

\item{software}{internal software to be used by Diffbind eg. edger or deseq2}

\item{genefile}{gene file name}

\item{path}{location of files}

\item{outformat}{eg. ".txt"}

\item{contrasts}{contrasts list}
}
\value{
list of annotated differential regions consensus
}
\description{
consensus peaks from nextflow output but can be applied to other sets
}
\examples{
annotation <- consensus_gene_count_anno("cutnrunko", "deseq2", cutnrunkd_deseq2_list$normalized_counts, cutnrunkd_deseq2_list$consensus, cutnrunkd_deseq2_list$gene, "/mnt/home3/reid/av638/cutnrun/iva_lab_oct23/cutnrun_k27_kd/outfolder/", "txt")
}
