% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{dds_counts}
\alias{dds_counts}
\title{deseq2 import raw counts}
\usage{
dds_counts(dds)
}
\arguments{
\item{dds}{DESeq2 dds object}
}
\value{
raw counts from dds object
}
\description{
deseq2 import raw counts
}
\examples{
counts <- dds_counts(rnaseqkd_list$dds)
}
