% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{deseq2_de}
\alias{deseq2_de}
\title{deseq2 prepare deseq2 matrix and perform differential analysis}
\usage{
deseq2_de(counts, coldata, fdr, fc)
}
\arguments{
\item{counts}{Count matrix (gene are rows and samples are columns and values are unnormalized)}

\item{coldata}{Meta data}

\item{fdr}{p.adj-value threshold}

\item{fc}{fold change threshold}
}
\value{
list (ALL, DE, UP and DOWN) of differentially expressed genes based on a given threshold
}
\description{
deseq2 prepare deseq2 matrix and perform differential analysis
}
\examples{
ddsde <- deseq2(rnaseqkd_list$processed_counts, rnaseqkd_list$coldata, 0.05, 2)
}
