% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{diffbind_and_packages_gene_anno}
\alias{diffbind_and_packages_gene_anno}
\title{diffbind annotation to genes and nearest distance filter for all packages on diffbind derived counts}
\usage{
diffbind_and_packages_gene_anno(
  assaytype,
  software,
  contrasts,
  genefile,
  path,
  outformat,
  fdr,
  fc
)
}
\arguments{
\item{assaytype}{eg. atacseq, rnaseq}

\item{software}{internal software to be used by Diffbind eg. edger or deseq2}

\item{contrasts}{contrasts list}

\item{genefile}{gene file name}

\item{path}{location of files}

\item{outformat}{eg. ".txt"}

\item{fdr}{p.adj-value threshold}

\item{fc}{fold change threshold}
}
\value{
list of annotated differential regions
}
\description{
diffbind annotation to genes and nearest distance filter for all packages on diffbind derived counts
}
\examples{
annotation <- diffbind_and_packages_gene_anno("atacseqkd", "diffbind_deseq2", atacseqkd_diffbind_deseq2_list$dar_analysis$de_analysis, atacseqkd_diffbind_deseq2_list$gene, "/mnt/home3/outfolder", "txt", 0.05, 2)
}
