% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{diffbind_count}
\alias{diffbind_count}
\title{diffbind get counts}
\usage{
diffbind_count(obj, summit)
}
\arguments{
\item{obj}{diffbind object}

\item{summit}{expected summit for peaks}
}
\value{
count matrix derived from diffbind
}
\description{
diffbind get counts
}
\examples{
counts <- diffbind_count(atacseqkd_diffbind_list$prep$dba_obj, 100)
}
