% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{diffbind_de}
\alias{diffbind_de}
\title{diffbind extract differential analysis results}
\usage{
diffbind_de(normobj, fdr, fc, method)
}
\arguments{
\item{normobj}{diffbind normalized object}

\item{fdr}{p.adj-value threshold}

\item{fc}{fold change threshold}

\item{method}{edgeR, deseq2}
}
\value{
differentially analysis outputs by diffbind
}
\description{
diffbind extract differential analysis results
}
\examples{
db_edgeR <- diffbind_de(atacseqkd_diffbind_list$norm$norm, 0.05, 2, "edgeR")
db_deseq2 <- diffbind_de(atacseqkd_diffbind_list$norm$norm, 0.05, 2, "deseq2")

}
