% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{diffbind_prep}
\alias{diffbind_prep}
\title{diffbind prepare object}
\usage{
diffbind_prep(samplesheet)
}
\arguments{
\item{samplesheet}{Samplesheet formatted as per Diffbind (see Diffbind on bioconductor for detail)}
}
\value{
samplesheet
}
\description{
diffbind prepare object
}
\examples{
ss <- diffbind_prep(diffbind_atacseqkd_samplesheet)
}
