% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{edger_create}
\alias{edger_create}
\title{edger create coldata/design, normalise, fit, mds plot}
\usage{
edger_create(data, coldata)
}
\arguments{
\item{coldata}{Meta data}

\item{counts}{Count matrix (gene are rows and samples are columns and values are unnormalized)}
}
\value{
prepare list of object required by edgeR
}
\description{
edger create coldata/design, normalise, fit, mds plot
}
\examples{
dge <- edger_create(atacseqkd_edger_list$processed_counts, atacseqkd_edger_list$coldata)
}
