% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{edger_de}
\alias{edger_de}
\title{edger differential analysis}
\usage{
edger_de(lmfit, design, fdr, fc)
}
\arguments{
\item{lmfit}{linear models fitted object prepared while performing edger_create()}

\item{design}{design matrices for specified variables}

\item{fdr}{p.adj-value threshold}

\item{fc}{fold change threshold}
}
\value{
list (ALL, DE, UP and DOWN) of differentially expressed genes based on a given threshold
}
\description{
edger differential analysis
}
\examples{
dar_analysis <- fn_edger_de(atacseqkd_edger_list$dge_obj$lmfit, atacseqkd_edger_list$dge_obj$design, 0.05, 2)
}
