% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{enrichment_counter}
\alias{enrichment_counter}
\title{enrichment_counter}
\usage{
enrichment_counter(list1, list2)
}
\arguments{
\item{list1}{peaklist1}

\item{list2}{peaklist2}
}
\value{
dataframe of overlap between two peak files
}
\description{
enrichment_counter
}
\examples{
df <- enrichment_counter(histonemarks_list1, atacseqkd_dars_list1)
}
