% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{integrate_bedtools}
\alias{integrate_bedtools}
\title{integrate multiple omics data on bedtools coverage output}
\usage{
integrate_bedtools(
  omics_data_path_lists,
  binsize,
  columns_pos,
  columns_samples
)
}
\arguments{
\item{omics_data_path_lists}{list of multiple outputs of bedtools coverage}

\item{binsize}{eg. "5kb}

\item{columns_pos}{eg. c(1:3)}

\item{columns_samples}{eg. vector of position of coverage column across all samples, can be extracted by seq(2,30,5) or manually c(2,7,12,17,22,27)}
}
\value{
dataframe for bedtools coverage performed on multiple samples
}
\description{
integrate multiple omics data on bedtools coverage output
}
\examples{
bin5kb <- integrate_bedtools(bin_path_combined, "5kb", c(1:3), seq(4,175,7))
}
