% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{integrate_featurecounts}
\alias{integrate_featurecounts}
\title{integrate multiple omics data on featurecounts}
\usage{
integrate_featurecounts(inputlist, sharedcolumn)
}
\arguments{
\item{inputlist}{list of dataframes from different samples /results}

\item{sharedcolumn}{eg. ensID_gene}
}
\value{
Make a consesnus based on given column presence in all samples eg. ensID_gene
}
\description{
integrate multiple omics data on featurecounts
}
\examples{
merge <- integrate_featurecounts(data_integration_list$ensID_gene$list, "ensID_Gene")
}
