% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{limma_create}
\alias{limma_create}
\title{limma create object}
\usage{
limma_create(data, coldata)
}
\arguments{
\item{data}{Count matrix}

\item{coldata}{Meta data}
}
\value{
Output a list of objects required for limma
}
\description{
limma create object
}
\examples{
dge <- limma_create(atacseqkd_limma_list$processed_counts, atacseqkd_limma_list$coldata)
}
