% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{limma_de}
\alias{limma_de}
\title{limma differential analysis}
\usage{
limma_de(lmfit, contrasts, fdr, fc)
}
\arguments{
\item{lmfit}{linear fitted model object}

\item{contrasts}{contarst for samples (use makeContrasts before running this function)}

\item{fdr}{p.adj-value threshold}

\item{fc}{fold change threshold}
}
\value{
list containing differential analysis results
}
\description{
limma differential analysis
}
\examples{
delist <- limma_de(atacseqkd_limma_list$dge_obj$fit, atacseqkd_limma_list$contrasts, 0.05, 2)
}
