% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{load_rdata}
\alias{load_rdata}
\title{deseq2 load rdata}
\usage{
load_rdata(path, rdata)
}
\arguments{
\item{path}{Location of the file}

\item{rdata}{RData object}
}
\value{
dds object
}
\description{
deseq2 load rdata
}
\examples{
dds <- load_rdata(rnaseqkd_path, "outfolder/star_salmon/deseq2_qc/deseq2.dds.RData")
}
