% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{make_coldata_dds}
\alias{make_coldata_dds}
\title{deseq2 coldata}
\usage{
make_coldata_dds(dds, samplefilter, rowsremove)
}
\arguments{
\item{dds}{DESEq2 dds object}

\item{samplefilter}{If any samples to filter otherwise skip}

\item{rowsremove}{Index of samples to filter}
}
\value{
coldata/ meta data df
}
\description{
deseq2 coldata
}
\examples{
coldata <- make_coldata_dds(rnaseqkd_list$dds, samplefilter = TRUE, c(7:9))
}
