% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{make_pca}
\alias{make_pca}
\title{pca prcomp using counts}
\usage{
make_pca(counts, file_extension, attribute)
}
\arguments{
\item{counts}{count matrix}

\item{file_extension}{bam file extension eg. ".bam"}

\item{attribute}{pca}
}
\value{
list of pca related objects and pca plot
}
\description{
pca prcomp using counts
}
\examples{
pca <- make_pca(cutntagwkd_quantify_list$featurecounts$featuresmatrix$bins5kb$bins5kb_countmatrix$hg38_5kb.txt$cpm, "_R1.target.markdup.sorted.bam", "pca")
}
