% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{maplot_general}
\alias{maplot_general}
\title{ggplot maplot}
\usage{
maplot_general(df, columnbyorder, fdr, logfc)
}
\arguments{
\item{df}{dataframe of output from deseq2 or other ways}

\item{columnbyorder}{since columns we renamed again the exact order need to be specified}

\item{fdr}{p.adj-value threshold}

\item{fc}{fold change threshold}
}
\value{
MA plot
}
\description{
ggplot maplot
}
\examples{
ma <- maplot_general(atacseqkd_limma_list$dar_analysis$dars$coef_shSUZ12_shControl$all, c(2,1,5,7), 0.05, 1)
}
