% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{maplot_general_shrinkage}
\alias{maplot_general_shrinkage}
\title{deseq2 maplot using ggplot2 using shirnkage}
\usage{
maplot_general_shrinkage(df, columnbyorder, fdr, logfc)
}
\arguments{
\item{df}{dataframe of output from deseq2 or other ways}

\item{columnbyorder}{since columns we renamed again the exact order need to be specified}

\item{fdr}{p.adj-value threshold}

\item{fc}{fold change threshold}
}
\value{
MA plot
}
\description{
deseq2 maplot using ggplot2 using shirnkage
}
\examples{
ma <- maplot_general_shrinkage(rnaseqkd_list$deseq2$de_analysis$shC1_c1509$all, c(1,3,6,7), 0.05, 1)
}
