% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{meta_plot}
\alias{meta_plot}
\title{ggplot meta plots}
\usage{
meta_plot(
  featurematrix,
  columnstoplot,
  multi = FALSE,
  value = "value",
  color = "col",
  linetype = "col",
  rep,
  splitside = 3,
  colorbyid = FALSE
)
}
\arguments{
\item{featurematrix}{count matrix}

\item{columnstoplot}{column with value to pick for plotting}

\item{multi}{FALSE}

\item{value}{column name with values to pick eg. "value",}

\item{color}{column name with col to pick  eg. "col"}

\item{linetype}{column name with col to pick  eg. "col"}

\item{rep}{eg. 1}

\item{splitside}{delimiter separator side 1,2,3,4,5,..... eg. 3}

\item{colorbyid}{FALSE}
}
\value{
Meta plots / Average plots / Density plot
}
\description{
ggplot meta plots
}
\examples{
meta <- meta_plot(atacseqkd_quantify_list$featurecounts$featuresmatrix, c(7:10), rep=1, splitside=3)
}
