% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{plot_venn}
\alias{plot_venn}
\title{deseq2 plot venn diagram}
\usage{
plot_venn(list, color)
}
\arguments{
\item{list}{list of samples (n < = 4) with vector of genes symbols}

\item{color}{vector of colors you want to give eg. c("black", "green", "red") for 3 samples}
}
\value{
venn diagram
}
\description{
deseq2 plot venn diagram
}
\examples{
venn <- plot_venn(list_rnaseqkd_de_shC1_c1509_shS_c1509, c("#0073C2FF", "#EFC000FF","#0073C2FF", "#EFC000FF"))
}
