% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{quantify_bams}
\alias{quantify_bams}
\title{Rsamtools quantify reads in a given bam files}
\usage{
quantify_bams(peaks_path, assaytype, bamfiles)
}
\arguments{
\item{peaks_path}{location of files}

\item{assaytype}{eg. cutandrun}

\item{bamfiles}{extension , eg. "\\.bam$"}
}
\value{
list of files with reads quantified
}
\description{
Rsamtools quantify reads in a given bam files
}
\examples{
bams <- quantify_bams("/mnt/home3/reid/av638/atacseq/iva_lab_gencode/boutfolder/bowtie2/merged_library/", "atacseqkd","\\\\.bam$")
}
