% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{quantify_featurecounts_multifeature_div}
\alias{quantify_featurecounts_multifeature_div}
\title{featurecounts quantify for multiple feature using featurecounts_div}
\usage{
quantify_featurecounts_multifeature_div(
  assay_bam_path,
  feature_path,
  assay,
  bam_extension,
  feature_list,
  columnstorearrange,
  ref,
  control,
  pe = FALSE,
  diff = "single",
  pairs
)
}
\arguments{
\item{assay_bam_path}{location of bam files}

\item{feature_path}{location of features files (eg. hg38_5kb.txt, gene_promoters.txt}

\item{assay}{eg. cutnandrun}

\item{bam_extension}{eg .bam}

\item{feature_list}{list of features to pick from the feature_path, all features shoudl have a specified arrangement of columns}

\item{columnstorearrange}{column rearrangement required}

\item{ref}{eg. hg38}

\item{control}{eg. IgG/Control}

\item{pe}{boolean {TRUE, FALSE}}

\item{diff}{eg. "single"}

\item{pairs}{list of sampleID, sample and respective control}
}
\value{
quantification in dataframe format for multiple sampples for multiple feature of interest
}
\description{
featurecounts quantify for multiple feature using featurecounts_div
}
\examples{
featuresmatrixdiv <- quantify_featurecounts_multifeature_div("/mnt/home3/reid/av638/cutntag/iva_lab_feb2024/bams/", "/mnt/home3/reid/av638/atacseq/iva_lab_gencode/integration/", "cutntagrmwt", "_sorted.bam",
cutntagrmwt_quantify_list$featurecounts$featureslist, c(4,5,1:3,6), "hg38", "IgG",pe=TRUE, diff="multi",
cutntagrmwt_quantify_list$featurecounts$pairs)
}
