% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{read_genefile}
\alias{read_genefile}
\title{import genome file}
\usage{
read_genefile(path, file)
}
\arguments{
\item{path}{Location of the file}

\item{file}{Gene file in a tab-separated format in order chr,start, end, strand, ensembl ID, gene symbol}
}
\value{
Gene file in a specified format
}
\description{
import genome file
}
\examples{
gene <- read_genefile(rnaseqkd_path, "gene_gencode_human__gencode_out_chr.txt")
}
