% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{rearrange_groups}
\alias{rearrange_groups}
\title{base rearrange groups}
\usage{
rearrange_groups(grouplist)
}
\arguments{
\item{grouplist}{list of path to specific genes(df) eg. cluster1.txt , cluster2.txt (prepared by list.files(integration_path, pattern="cluster*", full.names = T))}
}
\value{
filtered for random genes which you dont want in your data
}
\description{
base rearrange groups
}
\examples{
groups <- rearrange_groups(selected_rkd_gene_groups)
}
