% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{reassign_genenames}
\alias{reassign_genenames}
\title{deseq2 reassign gene names}
\usage{
reassign_genenames(countmatrix, genedf, columnskeep)
}
\arguments{
\item{countmatrix}{count matrix}

\item{genedf}{dataframe of gene coordinates and names}

\item{columnskeep}{columns you want to keep from the count matrix}
}
\value{
filtered count matrix
}
\description{
deseq2 reassign gene names
}
\examples{
count_mat <- reassign_genenames(rnaseqkd_list$samples_selected, rnaseqkd_list$gene, c(2:10))
}
