% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{replot_pie_chipseeker}
\alias{replot_pie_chipseeker}
\title{chipseeker pie plot}
\usage{
replot_pie_chipseeker(gr_anno)
}
\arguments{
\item{gr_anno}{granges annotation}
}
\value{
values for customising pie chart
}
\description{
chipseeker pie plot
}
\examples{
pie_val <- replot_pie_chipseeker(cutntagrmwt_quantify_list$chipseeker$plots$gr_anno)
}
