% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{run_chipseeker}
\alias{run_chipseeker}
\title{chipseeker run}
\usage{
run_chipseeker(peaks_path, assaytype, txdb, org_db, peakformat)
}
\arguments{
\item{peaks_path}{location of peaks files (eg. .bed , .narrowPeak, .broadPeak)}

\item{assaytype}{eg. atacseq}

\item{txdb}{eg. TxDb.Hsapiens.UCSC.hg38.knownGene (not quoted)}

\item{org_db}{eg. org.Hs.eg.db}

\item{peakformat}{eg. ".bed", all peak files should have same format}
}
\value{
list containing objects, granges, plots from chipseeker
}
\description{
chipseeker run
}
\examples{
list <- run_chipseeker("/mnt/home3/reid/av638/cutntag/iva_lab_feb2024/bams/endoH1peaks/", "cutntag", TxDb.Hsapiens.UCSC.hg38.knownGene, "org.Hs.eg.db", "*.broadPeak")
}
