% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{sample_filter}
\alias{sample_filter}
\title{deseq2 filter samples}
\usage{
sample_filter(countmatrix, samplefilter, columnsremove)
}
\arguments{
\item{countmatrix}{Gene expression raw counts matrix}

\item{samplefilter}{Sample to remove if not to be apart of analysis}

\item{columnsremove}{Index of those samples on column of datagrame}
}
\value{
Filtered raw count matrix selected for desired samples)
}
\description{
deseq2 filter samples
}
\examples{
filtered_count_matrix <- sample_filter(rnaseqkd_list$counts, samplefilter = TRUE, c(7:9))
}
