% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{scatterplot}
\alias{scatterplot}
\title{ggplot scatter}
\usage{
scatterplot(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
scatterplot
}
\description{
ggplot scatter
}
\examples{
sct <- scatterplot(cutntagrmwt_correlation_scatter_divlog2)
}
