% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{summedreads_per_feature}
\alias{summedreads_per_feature}
\title{Further process the featurecounts quantified matrix}
\usage{
summedreads_per_feature(fearturecountmatrix, bamcountsreads, labels)
}
\arguments{
\item{fearturecountmatrix}{Count matrix}

\item{bamcountsreads}{Dataframe of total reads in bam files}

\item{labels}{Labels of features short name eg. H3K4me3_1 K4me3}
}
\value{
Summed counts of reads in a given features
}
\description{
Further process the featurecounts quantified matrix
}
\examples{
summed <- summedreads_per_feature(atacseqkd_quantify_list$featurecounts$histonemarks$histone_marks_countmatrix, atacseqkd_quantify_list$bams$total_reads, atacseqkd_labelpath)
}
