% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwalker.R
\name{Upsetout}
\alias{Upsetout}
\title{Upset plot}
\usage{
Upsetout(input_list)
}
\arguments{
\item{input_list}{list of samples with vector of features eg. genes}
}
\value{
list of all possible intersections
}
\description{
Upset plot
}
\examples{
upset <- Upsetout(atacseqkd_venn_packages_list_CRAMP1)
}
